Refined logic for iOS and macOS
